Implemented everything except unit testing

+++ src/main/java/com/hexaware/maverickBank/service/interfaces/UserService.java
@@ -5,8 +5,8 @@ import com.hexaware.maverickBank.entity.User;
 public interface UserService {
     User registerUser(User user);
     User loginUser(String identifier, String password);
-    User getUserById(int userId);
-    User updateUser(int userId, String password, String email, String role, String status);
-    void deactivateUser(int userId);
-    void deleteUser(int userId);
+    User getUserById(Long userId);
+    User updateUser(Long userId, String password, String email, String role, String status);
+    void deactivateUser(Long userId);
+    void deleteUser(Long userId);
 }
